linux-io: File::into_raw_fd
This functionality was previously available only through an implementation
of a trait from the std crate. This new implementation is accessible even
in no_std mode.

This also fixes what was previously a bug in the IntoRawFd implementation,
where it would accidentally close the file descriptor by dropping the
File object before returning.