Wrapper functions return Result type
Previously we were making it the caller's responsibility to check whether
the raw result from the system calls was an error, but that didn't really
make sense in practice because the wrapper functions all return
specialized types that end up needing to be converted back to raw::V
immediately to do the check for error status.

Instead we'll build the error checks into the wrapper functions in a
similar way to what libc does, except of course we can return an error
code directly rather than relying on the global libc errno.

Once inlined the optimizer does a good job of reducing this to just a
branch based on a bitmask, so there doesn't seem to be any significant
mandatory overhead.