linux-io: Adapt to the new linux-unsafe wrapper API
The linux-unsafe wrappers now themselves return Result values, so we no
longer need to explicitly check for the error codes and wrap into a
Result in this package.

This crate does still have its own Error type because users of this crate
should not typically need to refer directly to linux-unsafe unless they
want to use one of the raw/unsafe APIs.