linux-unsafe: fcntl and ioctl wrappers
These ones are extra weird because the last argument's type and meaning
varies depending on the specified command/request.

We handle that here by exposing the previously-internal-only AsRawV type,
and then defining the final argument as impl AsRawV. Because these
functions are inlineable, in practice this should still reduce down to
just assigning a value directly to the relevant register once optimized.