linux-io: Safe interface to fcntl
The fcntl system call is tricky because it interprets its final argument
in very different ways depending on the given command value.

This abstraction adds type safety by annotating each of the command
constants with information about the argument type, which means that the
unsafe code is in the definition of the constants rather than in the uses
of those constants.