linux-io: SockAddrIp type for dual-stack IPv4/IPv6 applications
In real software we tend to need to support both IPv4 and IPv6 and choose
dynamically at runtime based on configuration or environment.

SockAddrIp and IpAddr together provide a light abstraction around the
protocol-specific types for situations where the protocol must be decided
dynamically. In the case of SockAddrIp the internal memory layout is still
compatible with passing a direct pointer to the payload into the system
call, and so this just adds the small extra step of checking the enum
discriminant value first to decide which address size to report.