linux-io: SockAddrIp more efficient memory layout
Previously we were using a Rust enum, which meant that it was possible to
use a normal exhaustive match but also meant wasting an extra byte to
store the tag.

The sockaddr type already has what is effectively a tag in the first
ushort field describing the address family, which is what the kernel
itself uses to recognize the address type. We can use the same field as
our userspace tag for a C-style union, thereby using exactly the same
layout as one would use in a C program.

This does mean that we now need to match based on the address family
constants rather than enum variants, but users of this dynamic address
type should typically not be directly inspecting its address family
anyway and instead just passing it through to the underlying socket API.