linux-io: ioctl device type checking
It isn't possible to expose a memory-safe interface to arbitrary ioctls
because the kernel has some ioctl request numbers that are overloaded with
different meanings in different drivers, and those different meanings can
have incompatible treatments of the arbitrary argument that may lead to
memory corruption if mismatched.

Therefore we'll track an optional device type for each File and expose
ioctl only for files that have an assigned device type. Assigning a device
type to a file is unsafe because the caller must somehow guarantee that
the file is indeed representing a device of the appropriate type, but once
that's done the actual use of the ioctl requests needs only safe Rust.