Fix various portability problems
The riscv port of Linux excludes a bunch of system functions that have been
made redundant by newer ones with superset functionality, so for linux-io
to be portable it must only use the modern functions.

This required adding a wrapper function for openat to unsafe, because
openat has subsumed open.