io: ioctl_writeread for write/read ioctl requests
We previously only supported one-way requests, but some ioctl requests
both write data to the kernel and then read data from the kernel, which
requires a slightly different strategy.

This also includes making the IoctlReq trait require Copy, because
otherwise it's messy to implement a restart-if-interrupted loop around
an interruptable ioctl request.