io: File::getdents_all
This is a more complete version of File::getdents that will call getdents64
repeatedly until the kernel reports that there are no more entries.

It isn't possible to implement Iterator over something that reuses buf
for multiple calls because earlier returned items can be invalidated by
subsequent calls to "next", and so here we use a trick of having the caller
provide a closure that transforms each DirEntry into something that can
be safely returned from the iterator, which therefore leaves the caller in
control of the policy for how that should occur.