linux-kvm: Explicitly set no-arg request args to zero
It seems that on some architectures the kernel is fussy about the argument
being set to zero specifically. Or perhaps it's fussy on all architectures
but so far I just got lucky and that register was always set to zero by
coincidence when I tested.

This introduces a new kind of IoctlReq that includes a const generic arg
specifying an integer to write into the argument, and then makes all of
the no-argument ioctl defs in linux-kvm use it.