linux-kvm: Better Debug impl for kvm_run
When we're using a nightly compiler we can use DebugStruct.field_with to
inline a custom presentation of the exit_details field that takes the
exit_reason value into account.

Hopefully this will get stabilized later so we can use it without an opt-in
but for now we have it behind a feature since linux-kvm otherwise does not
require a nightly compiler.